move record data to record.rs
